Fixed bugs with Clang in compiling busmezzo code
Code compiles now, but I have not checked results with test network.
TODO: compile with VS2012 to check if we broke anything.
Main changes:
1. “cosmetic” such as map <int, map<int, int>> to map <int, map<int,
int> > (note space)
2. destructors and ::reset() functions of Eventlist and Network now use
.clear() where possible, removed any iter=list.erase(iter) occurrences
3. added a To_String template to convert other types (double, int etc)
to string, as this is not in the clang version of std apparently
4. removed windows.h include in network.cpp and commented out its use
(error message dialog), as it is windows specific. replaced with a cout
output